requirements.txt
aiofiles==0.5.0
aiosqlite==0.13.0
appdirs==1.4.4
appnope==0.1.0
APScheduler==3.6.3
asgiref==3.2.7
backcall==0.2.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cookies==2.2.1
coverage==4.0.3
coveralls==1.1
cppyy-cling==6.20.4.1
dashboard==0.0.6
databases==0.3.2
decorator==4.0.9
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.13
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap3==12.1.0
django-facebook==6.0.3
django-image-cropping==1.4.0
django-imagekit==4.0.2
django-ipware==2.1.0
django-social-share==1.4.0
docopt==0.6.2
factory-boy==2.6.1
fake-factory==0.5.3
filelock==3.0.12
freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
funcsigs==0.4
gunicorn==19.4.5
h11==0.9.0
httplib2==0.9.2
idna==2.9
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mock==1.3.0
numpy==1.18.5
oauth2==1.9.0.post1
oauthlib==3.1.0
orm==0.2.0.dev1
parso==0.7.0
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.6
pilkit==2.0
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.5.1
Pygments==2.6.1
pyreadline==2.1
python-multipart==0.0.5
python-openid==2.2.5
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
responses==0.5.1
reverse==0.1.0
selenium==2.52.0
simplegeneric==0.8.1
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
starlette==0.13.4
static3==0.6.1
traitlets==4.1.0
typesystem==0.3.0.dev0
tzlocal==2.1
Unidecode==0.4.19
urllib3==1.25.9
uvicorn==0.11.5
virtualenv==20.0.21
wcwidth==0.2.4
websockets==8.1
whitenoise==5.1.0
